Applied commit 5c1cf0ee168c5dd604a83bb4f32a16faf8bf03b9 to branch v8-fixes (Changed type of Piece::IntType and nu_left_piece lists in Board from unsigned to uint_fast8_t.)

+++ src/libpentobi_base/Board.h
@@ -365,7 +365,7 @@ private:
             and restore_snapshot() */
         PiecesLeftList pieces_left;
 
-        PieceMap<unsigned> nu_left_piece;
+        PieceMap<uint_fast8_t> nu_left_piece;
 
         unsigned nu_onboard_pieces;
 
@@ -407,7 +407,7 @@ private:
     unsigned m_bonus_one_piece;
 
     /** See get_nu_piece_instances() */
-    unsigned m_nu_piece_instances;
+    uint_fast8_t m_nu_piece_instances;
 
     /** Same as m_board_const->get_move_info_array() */
     const MoveInfo* m_move_info_array;
